+++ hconfig_utils/mapl3g_hconfig_get.F90
@@ -1,0 +1,9 @@
+module mapl3g_hconfig_get
+
+   use mapl3g_hconfig_get_private, only: MAPL_HConfigGet => get_value
+
+   implicit none
+
+   public :: MAPL_HConfigGet
+
+end module mapl3g_hconfig_get